pyproject.toml
[tool.poetry]
name = "fastapicontactmanagement"
version = "0.1.0"
description = ""
authors = ["paulmusquaro <136118097+paulmusquaro@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2 = "^2.9.9"
sqlalchemy = "^2.0.30"
pydantic = {version = "^2.7.1", extras = ["email"]}
alembic = "^1.13.1"
fastapi = "^0.111.0"
uvicorn = {version = "0.24.0.post1", extras = ["standard"]}
python-jose = {version = "3.3.0", extras = ["cryptography"]}
libgravatar = "^1.0.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
mapped = "^20.4.2"
bcrypt = "^4.1.3"
fastapi-mail = "^1.4.1"
fastapi-limiter = "^0.1.6"
cloudinary = "^1.40.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
